# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAP API Documantation",
        "version": "latest"
    },
    "paths": {
...
        "/api/inventory": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Product List.",
                "operationId": "02191ace01f4f2e4a399a85886836a49",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "10"
                    },
                    {
                        "name": "get_all",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "json",
                                            "example": []
                                        },
                                        "links": {
                                            "type": "json",
                                            "example": []
                                        },
                                        "meta": {
                                            "type": "json",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "errors": {
                                            "type": "json",
                                            "example": {
                                                "message": [
                                                    "Unauthenticated"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    }
}